doc/examples/figures_paper1/fwdpy11_simulations/requirements.txt
numpy
tskit
demes
fwdpy11
moments-popgen
